Move `BackgroundDataStoreProcessor` to better namespace
